Merge lp:~smspillaz/compiz-core/compiz-core.use-constrainment into lp:compiz-core

(bzr r2943)

+++ src/window/constrainment/include/core/windowconstrainment.h
@@ -47,8 +47,8 @@ const unsigned int PHorzResizeInc = (1 << 1);
 
 CompSize constrainToHints (const XSizeHints &hints,
 			   const CompSize   &size,
-			   unsigned int     ignoreHints,
-			   unsigned int     resizeIgnoreHints);
+			   long ignoreHints,
+			   long resizeIgnoreHints);
 }
 
 }
